# licne-finansije/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Veb aplikacija za pregled licnih finansija",
        "description": "API za upravljanje ličnim finansijama (autentifikacija, transakcije, kategorije, budžeti, finansijski ciljevi, podsetnici, dokumenta i admin statistike)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
...
        "/api/podsetnici": {
            "get": {
                "tags": [
                    "Podsetnici"
                ],
                "summary": "Pregled svih podsetnika",
                "description": "Display a listing of the resource.",
                "operationId": "fb5d2c15cc2755e0ade3f3fe9194312c",
                "responses": {
                    "200": {
                        "description": "Lista svih podsetnika"
                    }
                }
            },
            "post": {
                "tags": [
                    "Podsetnici"
                ],
                "summary": "Kreiranje novog podsetnika",
                "description": "Store a newly created resource in storage.",
                "operationId": "3ded20b0a0ed5bd7d45a5dd61afdd958",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idKorisnik",
                                    "datum_vreme",
                                    "status"
                                ],
                                "properties": {
                                    "idKorisnik": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "opis": {
                                        "type": "string",
                                        "example": "Platiti struju"
                                    },
                                    "datum_vreme": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-06-20 10:00:00"
                                    },
                                    "status": {
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Podsetnik uspešno kreiran"
                    },
                    "422": {
                        "description": "Validacija nije prošla"
                    }
                }
            }
        },
        "/api/podsetnici/korisnik/{id}": {
            "get": {
                "tags": [
                    "Podsetnici"
                ],
                "summary": "Pregled podsetnika određenog korisnika",
                "operationId": "44e3b6af9f25c181bf78077ba515608f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID korisnika",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista podsetnika korisnika"
                    }
                }
            }
        },
        "/api/podsetnici/{id}": {
            "get": {
                "tags": [
                    "Podsetnici"
                ],
                "summary": "Prikaz detalja određenog podsetnika",
                "description": "Display the specified resource.",
                "operationId": "e7753a6b98ff1e821e583eaec3d5a041",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID podsetnika",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalji podsetnika"
                    },
                    "404": {
                        "description": "Podsetnik nije pronađen"
                    }
                }
            },
            "put": {
                "tags": [
                    "Podsetnici"
                ],
                "summary": "Ažuriranje postojećeg podsetnika",
                "description": "Update the specified resource in storage.",
                "operationId": "1bdc85e44dfeef728f6a0dc3642a8f96",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID podsetnika",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "idKorisnik": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "opis": {
                                        "type": "string",
                                        "example": "Promenjen opis podsetnika"
                                    },
                                    "datum_vreme": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-06-25 12:00:00"
                                    },
                                    "status": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Podsetnik uspešno ažuriran"
                    },
                    "422": {
                        "description": "Validacija nije prošla"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Podsetnici"
                ],
                "summary": "Brisanje podsetnika",
                "description": "Remove the specified resource from storage.",
                "operationId": "5bc195ac857ed7aa6f8b3f672cb3001b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID podsetnika",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Podsetnik je obrisan"
                    },
                    "404": {
                        "description": "Podsetnik nije pronađen"
                    }
                }
            }
        },
...
    },
...
}